Refactor code, fix attachment parsing error, enhance logging by message type, update dependencies, and bump package version to 1.6.29
