+++ package.json
@@ -1,19 +1,17 @@
 {
   "name": "@racla-dev/node-iris",
-  "version": "1.6.28",
+  "version": "1.6.29",
   "description": "TypeScript port of Python irispy-client module for KakaoTalk bot development",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
+    "build:tsc": "tsc",
     "watch": "tsc --watch",
     "clean": "rm -rf dist",
-    "test": "jest",
-    "test:watch": "jest --watch",
     "lint": "eslint src/**/*.ts",
     "format": "prettier --write src/**/*.ts",
-    "prepublishOnly": "npm run clean && npm run build && npm test",
-    "publish:public": "npm publish --access public"
+    "publish:public": "tsc && npm publish --access public"
   },
   "keywords": [
     "typescript",
@@ -44,31 +42,32 @@
   ],
   "devDependencies": {
     "@types/express": "^5.0.3",
-    "@types/jest": "^30.0.0",
-    "@types/node": "^22.18.0",
+    "@types/lodash": "^4.17.20",
+    "@types/module-alias": "^2.0.4",
+    "@types/node": "^22.18.3",
     "@types/node-cron": "^3.0.11",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.41.0",
-    "@typescript-eslint/parser": "^8.41.0",
+    "@types/ws": "^8.18.1",
+    "@typescript-eslint/eslint-plugin": "^8.43.0",
+    "@typescript-eslint/parser": "^8.43.0",
     "eslint": "^9.34.0",
-    "jest": "^30.0.5",
     "prettier": "^3.6.2",
-    "ts-jest": "^29.4.1",
     "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "^5.9.2"
   },
   "dependencies": {
-    "@types/lodash": "^4.17.20",
-    "@types/ws": "^8.5.12",
     "axios": "^1.11.0",
-    "dotenv": "^17.2.1",
+    "dotenv": "^17.2.2",
     "express": "^5.1.0",
     "got": "^11.8.6",
     "lodash": "^4.17.21",
+    "lossless-json": "^4.2.0",
+    "module-alias": "^2.2.3",
     "node-cron": "^4.2.1",
     "sharp": "^0.34.3",
     "tough-cookie": "^6.0.0",
-    "uuid": "^11.1.0",
+    "uuid": "^13.0.0",
     "winston": "^3.17.0",
     "ws": "^8.18.0"
   }
