+++ src/controllers/BootstrapController.ts
@@ -2,8 +2,8 @@
  * Bootstrap Controller for handling initialization tasks
  */
 
+import { Logger } from '@/utils/logger';
 import { BaseController } from './BaseController';
-import { Logger } from '../utils/logger';
 
 export class BootstrapController extends BaseController {
   private logger: Logger;
@@ -20,12 +20,12 @@ export class BootstrapController extends BaseController {
   // @Bootstrap(100) // 높은 우선순위로 먼저 실행
   // async initializeDatabase() {
   //   this.logger.info('Initializing database...');
-  //   
+  //
   //   try {
   //     // 데이터베이스 초기화 로직
   //     // 예: 스케줄된 메시지들을 DB에서 로드
   //     const scheduledMessages = await this.loadScheduledMessagesFromDB();
-  //     
+  //
   //     for (const msg of scheduledMessages) {
   //       this.scheduleMessage(
   //         msg.id,
@@ -35,7 +35,7 @@ export class BootstrapController extends BaseController {
   //         msg.metadata
   //       );
   //     }
-  //     
+  //
   //     this.logger.info(`Loaded ${scheduledMessages.length} scheduled messages from database`);
   //   } catch (error) {
   //     this.logger.error('Database initialization failed:', error);
@@ -45,7 +45,7 @@ export class BootstrapController extends BaseController {
   // @Bootstrap(50) // 중간 우선순위
   // async loadConfiguration() {
   //   this.logger.info('Loading configuration...');
-  //   
+  //
   //   try {
   //     // 설정 로드 로직
   //     this.logger.info('Configuration loaded successfully');
@@ -57,13 +57,15 @@ export class BootstrapController extends BaseController {
   /**
    * 데이터베이스에서 스케줄된 메시지를 로드하는 예제 메서드
    */
-  private async loadScheduledMessagesFromDB(): Promise<Array<{
-    id: string;
-    roomId: string;
-    message: string;
-    scheduledTime: number;
-    metadata?: any;
-  }>> {
+  private async loadScheduledMessagesFromDB(): Promise<
+    Array<{
+      id: string;
+      roomId: string;
+      message: string;
+      scheduledTime: number;
+      metadata?: any;
+    }>
+  > {
     // 실제 구현에서는 데이터베이스 연결 및 쿼리 수행
     // 예시로 빈 배열 반환
     return [];
