+++ src/controllers/ChatController.ts
@@ -2,9 +2,8 @@
  * Chat Controller for handling general chat events
  */
 
+import { Logger } from '@/utils/logger';
 import { BaseController } from './BaseController';
-import { ChatContext } from '../types/models';
-import { Logger } from '../utils/logger';
 
 export class ChatController extends BaseController {
   private logger: Logger;
