+++ src/controllers/DeleteMemberController.ts
@@ -3,9 +3,8 @@
  * Override methods in your bot implementation
  */
 
+import { Logger } from '@/utils/logger';
 import { BaseController } from './BaseController';
-import { ChatContext } from '../types/models';
-import { Logger } from '../utils/logger';
 
 export class DeleteMemberController extends BaseController {
   private logger: Logger;
